requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
pymysql
python-dotenv